add uri filter. move all rule init in derective commands

+++ src/ngx_http_tidehunter_debug.h
@@ -10,7 +10,7 @@
 #define IFDEBUG(block) do { if (DEBUG_FLAG) { block; } } while(0)
 
 #define PRINT_INFO(info) IFDEBUG( fprintf(stderr, info "\n") )
-#define PRINT_INT(info, i) IFDEBUG( fprintf(stderr, info "%d\n", i) )
+#define PRINT_INT(info, i) IFDEBUG( fprintf(stderr, info "%d\n", (int)i) )
 #define PRINT_FLOAT(info, f) IFDEBUG( fprintf(stderr, info "%f\n", f) )
 #define PRINT_STR(info, str) IFDEBUG( fprintf(stderr, info " #%s#\n", str) )
 #define PRINT_NGXSTR(info, nstr) IFDEBUG( fprintf(stderr, info " #%.*s#\n", (int)nstr.len, nstr.data) )
